requirements.txt
beautifulsoup4==4.10.0
black==22.1.0
blis==0.7.5
boto3==1.20.47
botocore==1.23.47
cached-path==1.0.2
cachetools==5.0.0
catalogue==2.0.6
chardet==4.0.0
charset-normalizer==2.0.11
checklist==0.0.11
cheroot==8.6.0
cherrypy==18.6.1
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
cymem==2.0.6
datasets==1.18.3
debugpy==1.6.0
dill==0.3.4
docker-pycreds==0.4.0
einops==0.4.0
entrypoints==0.4
executing==0.8.2
fairscale==0.4.5
feedparser==6.0.8
filelock==3.4.2
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
gitpython==3.1.26
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
h5py==3.6.0
higher==0.2.1
huggingface-hub==0.2.1
hydra-core==1.1.1
idna==3.3
iniconfig==1.1.1
ipykernel==6.9.2
ipython==8.0.1
ipywidgets==7.7.0
iso-639==0.4.5
jaraco-classes==3.2.1
jaraco-collections==3.5.1
jaraco-context==4.1.1
jaraco-functools==3.5.0
jaraco-text==3.7.0
jinja2==3.1.1
jmespath==0.10.0
joblib==1.1.0
jsonnet==0.18.0
jsonschema==4.4.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
langcodes==3.3.0
lmdb==1.3.0
lxml==4.7.1
markupsafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.12.0
multidict==6.0.2
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==1.0.6
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.1
nbformat==5.2.0
nest-asyncio==1.5.4
nltk==3.6.5
notebook==6.4.10
numpy==1.22.1
omegaconf==2.1.1
pandas==1.4.0
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.1
patternfork-nosql==3.6
pdfminer-six==20211012
pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
portend==3.1.0
preshed==3.0.6
promise==2.3
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
pyyaml==6.0
qtpy==2.0.1
regex==2022.1.18
requests==2.27.1
rsa==4.8
s3transfer==0.5.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
sentry-sdk==1.5.4
sgmllib3k==1.0.0
shortuuid==1.0.8
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.1.4
tempora==5.0.1
tensorboardx==2.4.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.0
torchvision==0.11.3
tqdm==4.62.3
transformers==4.15.0
typer==0.4.0
urllib3==1.26.8
wandb==0.12.10
wasabi==0.9.0
webencodings==0.5.1
widgetsnbextension==3.6.0
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
zc-lockfile==2.0